grant: no level up for grant
